requirements.txt
spacy==3.2.0
nltk==3.6.2
pyresparser==1.0.6
python-dotenv
openai==1.3.3
PyPDF2
resume-parser